Extract stories out of stories folder
