+++ dashboard/src/components/common/Header/Header.stories.tsx
@@ -1,5 +1,4 @@
-import React from "react"
-import { ComponentStory, ComponentMeta } from "@storybook/react"
+import { ComponentMeta, ComponentStory } from "@storybook/react"
 
 import { Header } from "./Header"
