+++ dashboard/src/components/common/Header/Header.tsx
@@ -1,17 +1,17 @@
-import React from "react"
+import React from "react";
 
-import { Button } from "./Button"
-import "./header.css"
+import "./header.css";
+import { Button } from "../Button/Button";
 
 type User = {
-  name: string
-}
+  name: string;
+};
 
 interface HeaderProps {
-  user?: User
-  onLogin: () => void
-  onLogout: () => void
-  onCreateAccount: () => void
+  user?: User;
+  onLogin: () => void;
+  onLogout: () => void;
+  onCreateAccount: () => void;
 }
 
 export const Header = ({
@@ -68,4 +68,4 @@ export const Header = ({
       </div>
     </div>
   </header>
-)
+);
