+++ dashboard/src/components/common/Page/Page.tsx
@@ -1,14 +1,14 @@
-import React from "react"
+import React from "react";
 
-import { Header } from "./Header"
-import "./page.css"
+import "./page.css";
+import { Header } from "../Header/Header";
 
 type User = {
-  name: string
-}
+  name: string;
+};
 
 export const Page: React.VFC = () => {
-  const [user, setUser] = React.useState<User>()
+  const [user, setUser] = React.useState<User>();
 
   return (
     <article>
@@ -87,5 +87,5 @@ export const Page: React.VFC = () => {
         </div>
       </section>
     </article>
-  )
-}
+  );
+};
